Refactor pkg/bot-service to pkg/service
ARI-53
