+++ slackbot/cmd/bot-server/main.go
@@ -2,8 +2,8 @@ package main
 
 import (
 	"fmt"
-	"github.com/buildit/slackbot/pkg/bot-server"
 	"github.com/buildit/slackbot/pkg/database"
+	"github.com/buildit/slackbot/pkg/service"
 	"github.com/gorilla/mux"
 	"log"
 	"net/http"
@@ -19,10 +19,10 @@ func main() {
 	}
 
 	router := mux.NewRouter()
-	router.HandleFunc("/", bot_server.ListenAndServeHome)
-	router.HandleFunc("/events", bot_server.ListenAndServeEvents).Methods("POST")
-	router.HandleFunc("/slash", bot_server.ListenAndServeSlash).Methods("POST")
-	router.HandleFunc("/interactions", bot_server.ListenAndServeInteractions).Methods("POST")
+	router.HandleFunc("/", service.ListenAndServeHome)
+	router.HandleFunc("/events", service.ListenAndServeEvents).Methods("POST")
+	router.HandleFunc("/slash", service.ListenAndServeSlash).Methods("POST")
+	router.HandleFunc("/interactions", service.ListenAndServeInteractions).Methods("POST")
 	fmt.Println("[INFO] Server listening")
 	log.Fatal(http.ListenAndServe(":80", router))
 	defer database.CloseDB()
